requirements.txt
# conda create -n tf-win python=3.9 -y
# conda activate tf-win

# pip install -r requirements.txt

# 有 NVIDIA 显卡 (推荐)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# 没有独立显卡 (使用 CPU)
# pip install torch torchvision torchaudio

# 防止 NumPy 被自动升级
# pip install "numpy==1.24.3"

# torch
# torchvision
# torchaudio

tensorflow==2.13.0
numpy==1.24.3
typing-extensions==4.5.0
pandas==2.0.3
scikit-learn==1.3.2
scipy
opencv-python<4.10
Pillow
mysql-connector-python
joblib
tqdm
matplotlib
seaborn